+++ luakeys.lua
@@ -265,54 +265,6 @@ local utils = (function()
     throw_error_message(message, help)
   end
 
-  ---
-  ---Scan for an optional argument.
-  ---
-  ---@param initial_delimiter? string # The character that marks the beginning of an optional argument (by default `[`).
-  ---@param end_delimiter? string # The character that marks the end of an optional argument (by default `]`).
-  ---
-  ---@return string|nil # The string that was enclosed by the delimiters. The delimiters themselves are not returned.
-  local function scan_oarg(initial_delimiter,
-    end_delimiter)
-    if initial_delimiter == nil then
-      initial_delimiter = '['
-    end
-
-    if end_delimiter == nil then
-      end_delimiter = ']'
-    end
-
-    ---
-    ---@param t Token
-    ---
-    ---@return string
-    local function convert_token(t)
-      if t.index ~= nil then
-        return utf8.char(t.index)
-      else
-        return '\\' .. t.csname
-      end
-    end
-
-    local function get_next_char()
-      local t = token.get_next()
-      return convert_token(t), t
-    end
-
-    local char, t = get_next_char()
-    if char == initial_delimiter then
-      local oarg = {}
-      char = get_next_char()
-      while char ~= end_delimiter do
-        table.insert(oarg, char)
-        char = get_next_char()
-      end
-      return table.concat(oarg, '')
-    else
-      token.put_next(t)
-    end
-  end
-
   local function visit_tree(tree, callback_func)
     if type(tree) ~= 'table' then
       throw_error_message(
@@ -698,7 +650,6 @@ local utils = (function()
     tex_printf = tex_printf,
     throw_error_message = throw_error_message,
     throw_error_code = throw_error_code,
-    scan_oarg = scan_oarg,
     ansi_color = ansi_color,
     log = log,
   }