+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+### Removed
+
+- Remove helper function `utils.scan_oarg()`, use [lparse](https://www.ctan.org/pkg/lparse) instead.
+
 ## [0.13.0] - 2023-01-13
 
 ### Added